# Remove superfluous timing transformations.