Before:
``` before:
- finished in 62.6510771s, task avg 1ms, min 1ms, max 125ms
- finished in 54.6367758s, task avg 1ms, min 1ms, max 50ms
- finished in 54.2207575s, task avg 1ms, min 1ms, max 32ms

 ```
After:
``` after: 
- finished in 53.5315927s, task avg 1.824454ms, min 1.6432ms, max 37.6035ms
- finished in 54.3072565s, task avg 1.846359ms, min 1.6451ms, max 46.814ms
- finished in 53.7990169s, task avg 1.834777ms, min 1.6462ms, max 36.9701ms

```

We needed to fix this so we can properly measure and see the future changes.